requirements.txt
alembic==1.0.11
astroid==2.2.5
Click==7.0
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.0
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gunicorn==19.9.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
Mako==1.0.14
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.18.2
psycopg2==2.8.3
pylint==2.3.1
pymongo==3.8.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.6
typed-ast==1.4.0
Werkzeug==0.15.5
wrapt==1.11.2
WTForms==2.2.1
